set props
